requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.6
paho-mqtt==2.1.0
Flask-login==0.6.3